+++ tests/Integration/Request/CreateUserTest.php
@@ -45,22 +45,6 @@ final class CreateUserTest extends IntegrationTestCase
         $this->auth->deleteUser($user->uid);
     }
 
-    public function testCreateUserWithoutEmailButMarkTheEmailAsVerified(): void
-    {
-        $request = CreateUser::new()
-            ->withUid($uid = \bin2hex(\random_bytes(5)))
-            ->markEmailAsVerified()
-        ;
-
-        $user = $this->auth->createUser($request);
-
-        $this->assertSame($uid, $user->uid);
-        $this->assertNull($user->email);
-        $this->assertFalse($user->emailVerified);
-
-        $this->auth->deleteUser($user->uid);
-    }
-
     public function testCreateUserWithoutEmailButMarkTheEmailAsUnverified(): void
     {
         $request = CreateUser::new()